requirements.txt
Django>=6.0,<7.0
django_components>=0.148
Pillow>=12.0
polib>=1.2
gunicorn>=23.0
whitenoise>=6.0
